requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
python-dotenv==1.0.0
deep-translator==1.11.4
langdetect==1.0.9
google-generativeai==0.3.2
requests==2.31.0
markdown==3.5.1
gunicorn==21.2.0
textblob==0.17.1
werkzeug==2.3.7
